Add node GetConfigHash implementation
